move display profile pref
