+++ app/src/main/java/eu/kanade/tachiyomi/ui/reader/ReaderActivity.kt
@@ -816,7 +816,7 @@ class ReaderActivity : BaseActivity() {
                 }
                 .launchIn(lifecycleScope)
 
-            readerPreferences.displayProfile().changes()
+            preferences.displayProfile().changes()
                 .onEach { setDisplayProfile(it) }
                 .launchIn(lifecycleScope)
